Fix static analysis warnings
#rb trivial
#jira none

[CL 30999195 by Luke Thatcher in ue5-main branch]

+++ Engine/Source/Runtime/RHI/Private/RHICommandList.cpp
@@ -663,7 +663,7 @@ void FRHICommandListExecutor::FSubmitState::Dispatch(FRHICommandListBase* CmdLis
 			) || (
 				MaxCommandsPerTranslate  < 0 ? true  :
 				MaxCommandsPerTranslate == 0 ? false :
-				(CurrentTranslateJob->NumCommands + CmdList->NumCommands) > MaxCommandsPerTranslate
+				(CurrentTranslateJob->NumCommands + CmdList->NumCommands) > uint32(MaxCommandsPerTranslate)
 			))
 		{
 			FinalizeCurrent();
